Allow to pass a config in a backwards compatible way to plugins
Check the signature of the precommit function before calling it. If the
precommit function has arguments it's safe to pass the config dict as
keyword argument. Otherwise run the precommit function without an
argument and print a deprecation warning. The support for a precommit
function without arguments should be removed with autohooks 2.0.

autohooks.plugins.pylint and autohooks.plugins.black will get an update
soon.