Update setup.py to read most settings from pyproject.toml
Use the poetry config from the pyproject.toml file to set most setting
in setup.py.

We are still using setup.py instead of poetry as a build-system (PEP 517)
because setup.py allows us to activate the git hooks as a post install
when beeing installed via a source distribution (.tar.gz file).